make main more generic
